requirements.txt
pandas>=2.1
pyarrow>=15.0
fastparquet>=2024.2.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
tqdm>=4.66
python-dateutil>=2.9
jinja2>=3.1
matplotlib>=3.8
seaborn>=0.13
plotly>=5.20
tabulate>=0.9
fastapi>=0.111
uvicorn>=0.30
PyYAML>=6.0.2
streamlit>=1.32
openpyxl>=3.1
xlrd==1.2.0
